code duplication refactor
